requirement.txt
django
djangorestframework
anvirons
mysqlclient